requirements.txt
diffusers==0.29.2
transformers==4.38.2
accelerate>=0.27,<0.28
safetensors
einops
omegaconf
huggingface-hub
opencv-python
aiohttp
pyyaml
Pillow
importlib_metadata
numpy
mediapipe